+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- Hi! -->
 
+## tdl (unreleased)
+
+- Added `tdl.setLogMessageCallback` that allows to pass a callback to the
+  `td_set_log_message_callback` TDLib function using Node-API's thread-safe
+  functions. (TDLib v1.8.0+ only)
+
 ## tdl-install-types@0.1.0 (2023-09-26)
 
 Initial release of a new cli utility that is now recommended to use instead of